i2c slave updates

+++ embassy-stm32/src/can/fd/message_ram/txevent_element.rs
@@ -88,7 +88,7 @@ impl R {
     #[doc = "Byte 0 - Bits 0:28 - ID"]
     #[inline(always)]
     pub fn id(&self) -> ID_R {
-        ID_R::new((self.bits[0]) & 0x1FFFFFFF)
+        ID_R::new(((self.bits[0]) & 0x1FFFFFFF) as u32)
     }
     #[doc = "Byte 0 - Bit 29 - RTR"]
     #[inline(always)]
